Interactive API: Refactor/update metadata interfaces
[#172778790]
[#172778934]
