+++ docs/interactive-api-client/interfaces/iauthoringinteractivemetadata.md
@@ -2,6 +2,8 @@
 
 # Interface: IAuthoringInteractiveMetadata
 
+Generic interactive type.
+
 ## Hierarchy
 
 * [IAuthoringMetadataBase](iauthoringmetadatabase.md)
@@ -12,20 +14,12 @@
 
 ### Properties
 
-* [isRequired](iauthoringinteractivemetadata.md#isrequired)
 * [prompt](iauthoringinteractivemetadata.md#optional-prompt)
-* [secondaryTypeForNow](iauthoringinteractivemetadata.md#secondarytypefornow)
-* [type](iauthoringinteractivemetadata.md#type)
+* [questionType](iauthoringinteractivemetadata.md#questiontype)
+* [required](iauthoringinteractivemetadata.md#optional-required)
+* [subtype](iauthoringinteractivemetadata.md#optional-subtype)
 
 ## Properties
-
-###  isRequired
-
-• **isRequired**: *boolean*
-
-*Inherited from [IAuthoringMetadataBase](iauthoringmetadatabase.md).[isRequired](iauthoringmetadatabase.md#isrequired)*
-
-___
 
 ### `Optional` prompt
 
@@ -35,14 +29,27 @@ ___
 
 ___
 
-###  secondaryTypeForNow
+###  questionType
 
-• **secondaryTypeForNow**: *any*
+• **questionType**: *"iframe_interactive"*
 
-*Inherited from [IAuthoringMetadataBase](iauthoringmetadatabase.md).[secondaryTypeForNow](iauthoringmetadatabase.md#secondarytypefornow)*
+*Overrides [IAuthoringMetadataBase](iauthoringmetadatabase.md).[questionType](iauthoringmetadatabase.md#questiontype)*
 
 ___
 
-###  type
+### `Optional` required
 
-• **type**: *"interactive"*
+• **required**? : *undefined | false | true*
+
+*Inherited from [IAuthoringMetadataBase](iauthoringmetadatabase.md).[required](iauthoringmetadatabase.md#optional-required)*
+
+___
+
+### `Optional` subtype
+
+• **subtype**? : *undefined | string*
+
+*Inherited from [IAuthoringMetadataBase](iauthoringmetadatabase.md).[subtype](iauthoringmetadatabase.md#optional-subtype)*
+
+Interactive can optionally set subtype, so Teacher Report can display different icons
+or categorize interactives into subcategories.
