+++ lara-typescript/src/interactive-api-client/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
+export * from "./metadata-types";
 export * from "./in-frame";
 export * from "./api";
 export * from "./hooks";
